+++ model/autosklearn/dffml_model_autosklearn/config.py
@@ -4,7 +4,6 @@ from typing import Optional, AsyncIterator, List, Dict, Any
 
 import pandas as pd
 import joblib as joblib
-import sklearn as sklearn
 
 from dffml.base import config, field
 from dffml.model.accuracy import Accuracy
@@ -139,9 +138,6 @@ class AutoSklearnModelContext(ModelContext):
     async def get_predictions(self, data):
         return self.model.predict(data)
 
-    async def get_probabilities(self, data):
-        return self.model.predict_proba(data)
-
     async def train(self, sources: Sources):
         all_data = []
         async for record in sources.with_features(
@@ -187,7 +183,7 @@ class AutoSklearnModelContext(ModelContext):
         y_test = df[[self.parent.config.predict.name]]
         x_test = df.drop(columns=[self.parent.config.predict.name])
         predictions = await self.get_predictions(x_test)
-        accuracy = sklearn.metrics.accuracy_score(y_test, predictions)
+        accuracy = await self.accuracy_score(y_test, predictions)
         return Accuracy(accuracy)
 
     @property